refactor: move channel channelService to Channel
