+++ src/Resources/config/services.xml
@@ -22,6 +22,10 @@
         </service>
 
         <!-- core service -->
+        <service id="ems.service.channel" class="EMS\CoreBundle\Service\Channel\ChannelService">
+            <argument type="service" id="ems.repository.channel"/>
+            <argument type="service" id="logger"/>
+        </service>
         <service id="ems.service.channel.register" class="EMS\CoreBundle\Service\Channel\ChannelRegistrar">
             <argument type="service" id="ems.repository.channel"/>
             <argument type="service" id="emsch.helper_environment"/>
